get all recent changes

+++ CMR/src/info/novatec/inspectit/cmr/service/SecurityService.java
@@ -1,6 +1,10 @@
 package info.novatec.inspectit.cmr.service;
 
+import java.security.KeyPair;
+import java.security.KeyPairGenerator;
+import java.security.NoSuchAlgorithmException;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 import javax.annotation.PostConstruct;
@@ -8,6 +12,8 @@ import javax.annotation.PostConstruct;
 import org.apache.shiro.SecurityUtils;
 import org.apache.shiro.authc.AuthenticationException;
 import org.apache.shiro.authc.UsernamePasswordToken;
+import org.apache.shiro.session.Session;
+import org.apache.shiro.session.mgt.DefaultSessionManager;
 import org.apache.shiro.subject.Subject;
 import org.slf4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,16 +30,18 @@ import info.novatec.inspectit.communication.data.cmr.Permutation;
 import info.novatec.inspectit.communication.data.cmr.Role;
 import info.novatec.inspectit.communication.data.cmr.User;
 import info.novatec.inspectit.spring.logger.Log;
+import info.novatec.inspectit.util.PermutationException;
 
 /**
- * Provides general security-system operations for client<->cmr interaction. Watches over Data
- * Integrity.
+ * Provides general security-system operations for client<->cmr interaction.
+ * Watches over Data Integrity.
  * 
  * @author Andreas Herzog
  * @author Clemens Geibel
  * @author Lucca Hellriegel
  * @author Mario Rose
  * @author Joshua Hartmann
+ * @author Phil Szalay
  */
 @Service
 public class SecurityService implements ISecurityService {
@@ -60,19 +68,31 @@ public class SecurityService implements ISecurityService {
 	 * Manager for general security purposes.
 	 */
 	@Autowired
-	CmrSecurityManager cmrSecurityManager;
+	CmrSecurityManager securityManager;
 
 	/**
 	 * Data Access Object.
 	 */
 	@Autowired
 	RoleDao roleDao;
+	
+	/**
+	 * KeyPair.
+	 */
+	private KeyPair keyPair;
 
 	/**
-	 * Is executed after dependency injection is done to perform any initialization.
+	 * Is executed after dependency injection is done to perform any
+	 * initialization.
 	 */
 	@PostConstruct
 	public void postConstruct() {
+		try {
+			setKeyPair();
+		} catch (NoSuchAlgorithmException nsaEx) {
+			log.info(nsaEx.getMessage());
+		}
+		
 		if (log.isInfoEnabled()) {
 			log.info("|-Security Service active...");
 		}
@@ -83,23 +103,26 @@ public class SecurityService implements ISecurityService {
 	// +-------------------------------------------------------------------------------------------+
 
 	/**
-	 * Authentication via the CmrSecurityManager.
-	 * 
-	 * @param pw
-	 *            users password
-	 * @param email
-	 *            email
-	 * @return whether the login was successful
+	 * {@inheritDoc}
 	 */
 	@Override
-	public boolean authenticate(String pw, String email) {
+	public boolean authenticate(byte[] encryptedRandomKey, byte[] secondEncryptionLevel, String email) {
+		String pw;
+		try {
+			pw = Permutation.decryptPassword(encryptedRandomKey, secondEncryptionLevel, keyPair.getPrivate().getEncoded());
+		} catch (Exception e) {
+			log.info(e.getMessage());
+			return false;
+		}
+		
 		UsernamePasswordToken token = new UsernamePasswordToken(email, pw);
 
 		Subject currentUser = SecurityUtils.getSubject();
-		
-		if (userDao.findByEmail(email).isLocked()) { 
-			return false; 
-			}
+
+		User user = userDao.findByEmail(email);
+		if (user == null || user.isLocked()) {
+			return false;
+		}
 
 		if (!currentUser.isAuthenticated()) {
 			try {
@@ -115,6 +138,29 @@ public class SecurityService implements ISecurityService {
 
 		return true;
 	}
+	
+	
+	/**
+	 * First Step of the modified login process.
+	 * @param secretKeyBytes symmetricKey
+	 * @return symmetrically encrypted public key
+	 * @throws PermutationException 
+	 */
+	@Override
+	public byte[] callPublicKey(byte[] symmetricKey) throws PermutationException {
+		byte[] encryptedPublicKey = Permutation.encryptPublicKey(keyPair.getPublic(), symmetricKey);
+		return encryptedPublicKey;
+	}
+	
+	/**
+	 * Initializes the public/private keys.
+	 * @throws NoSuchAlgorithmException if RSA not available.
+	 */
+	private void setKeyPair() throws NoSuchAlgorithmException {
+		KeyPairGenerator generator = KeyPairGenerator.getInstance(Permutation.ASYMMETRIC_ALGORITHM);
+		generator.initialize(Permutation.ASYMMETRIC_KEY_SIZE);
+	    this.keyPair = generator.generateKeyPair();
+	}
 
 	/**
 	 * Ends the session.
@@ -123,7 +169,7 @@ public class SecurityService implements ISecurityService {
 	public void logout() {
 		SecurityUtils.getSubject().logout();
 	}
-	
+
 	/**
 	 * Returns whether the user is authenticated.
 	 * 
@@ -134,14 +180,14 @@ public class SecurityService implements ISecurityService {
 	}
 
 	/**
-	 * Returns titles of permissions as Strings.
+	 * Returns list of permissions.
 	 * 
 	 * @return List with the users permissions.
 	 */
 	@Override
 	public List<Permission> getPermissions() {
 		Subject currentUser = SecurityUtils.getSubject();
-		
+
 		List<Permission> grantedPermissions = new ArrayList<Permission>();
 		List<Permission> existingPermissions = permissionDao.loadAll();
 		for (int i = 0; i < existingPermissions.size(); i++) {
@@ -158,8 +204,8 @@ public class SecurityService implements ISecurityService {
 	// +-------------------------------------------------------------------------------------------+
 
 	/**
-	 * Combines the integrity check for all security data types. Uniqueness etc. is specifically
-	 * checked in every method.
+	 * Combines the integrity check for all security data types. Uniqueness etc.
+	 * is specifically checked in every method.
 	 * 
 	 * @param data
 	 *            data
@@ -173,16 +219,20 @@ public class SecurityService implements ISecurityService {
 			Permission permission = (Permission) data;
 			return (permission.getDescription().length() < 100);
 		} else if (data instanceof Role) {
-			//TODO: make real data integrity tests
+			// TODO: make real data integrity tests
 			return true;
 		}
-		
+
 		return false;
 	}
 
 	// | USER |---------------
 	@Override
 	public List<String> getAllUsers() {
+		if (!securityManager.isPermitted("cmrAdministrationPermission")) {
+			return new ArrayList<String>();
+		}
+		
 		List<User> users = userDao.loadAll();
 		List<String> userEmails = new ArrayList<String>();
 		for (User user : users) {
@@ -191,9 +241,13 @@ public class SecurityService implements ISecurityService {
 
 		return userEmails;
 	}
-	
+
 	@Override
 	public List<String> getUsersByRole(long id) {
+		if (!securityManager.isPermitted("cmrAdministrationPermission")) {
+			return new ArrayList<String>();
+		}
+		
 		List<User> foundUsers = userDao.findByRole(id);
 		List<String> userEmails = new ArrayList<String>();
 		for (User user : foundUsers) {
@@ -205,6 +259,10 @@ public class SecurityService implements ISecurityService {
 
 	@Override
 	public void addUser(User user) throws DataIntegrityViolationException {
+		if (!securityManager.isPermitted("cmrAdministrationPermission")) {
+			return;
+		}
+		
 		if (!checkDataIntegrity(user)) {
 			throw new DataIntegrityViolationException("Data integrity test failed!");
 		}
@@ -217,17 +275,25 @@ public class SecurityService implements ISecurityService {
 		} else {
 			String hashedPassword = Permutation.hashString(user.getPassword());
 			user.setPassword(hashedPassword);
-			userDao.saveOrUpdate(user);					
+			userDao.saveOrUpdate(user);
 		}
 	}
 
 	@Override
 	public User getUser(String email) {
+		if (!securityManager.isPermitted("cmrAdministrationPermission")) {
+			return null;
+		}
+		
 		return userDao.findByEmail(email);
 	}
 
 	@Override
 	public void deleteUser(User user) {
+		if (!securityManager.isPermitted("cmrAdministrationPermission")) {
+			return;
+		}
+		
 		Subject currentUser = SecurityUtils.getSubject();
 		String currentName = (String) currentUser.getPrincipal();
 		if (currentName.equals(user.getEmail())) {
@@ -236,58 +302,85 @@ public class SecurityService implements ISecurityService {
 		userDao.delete(user);
 	}
 
-	//TODO: TESTMETHODE!
 	@Override
-	public void changeUserAttribute(User userOld, String email, String password, long roleID, boolean passwordChanged, boolean isLocked) {
+	public void changeUserAttribute(User userOld, String email, String password, long roleID, boolean passwordChanged,
+			boolean isLocked) {
+		if (!securityManager.isPermitted("cmrAdministrationPermission")) {
+			return;
+		}
+		
 		Subject currentUser = SecurityUtils.getSubject();
 		String currentName = (String) currentUser.getPrincipal();
 		if (currentName.equals(userOld.getEmail())) {
 			currentUser.logout();
 		}
-		if (passwordChanged) {
-			User userNew = new User(password, email, roleID, isLocked);
-			userDao.delete(userOld);
-			addUser(userNew);
-		} else {
-			User userNew = new User(userOld.getPassword(), email, roleID, isLocked);
-			userDao.delete(userOld);
-			userDao.saveOrUpdate(userNew); //this way the old password is not hashed twice.
+
+		if (!email.equals(userOld.getEmail()) && userDao.findByEmail(email) != null) {
+			throw new DataIntegrityViolationException("User with this email does already exist!");
 		}
+		if (roleDao.findByID(roleID) == null) {
+			throw new DataIntegrityViolationException("Invalid role id assigned to this user!");
+		}
+
+		userOld.setEmail(email);
+		userOld.setRoleId(roleID);
+		userOld.setLocked(isLocked);
+		if (passwordChanged) {
+			String hashedPassword = Permutation.hashString(password);
+			userOld.setPassword(hashedPassword);
+		}
+		if (!checkDataIntegrity(userOld)) {
+			throw new DataIntegrityViolationException("Data integrity test failed!");
+		}
+		userDao.saveOrUpdate(userOld);
 	}
 
 	// | PERMISSION |---------
-	
+
 	@Override
 	public void changePermissionDescription(Permission permission) {
-		changePermissionAttributes(permission, permission.getTitle(), permission.getDescription(), permission.getParameter());
+		if (!securityManager.isPermitted("cmrAdministrationPermission")) {
+			return;
+		}
+		
+		changePermissionAttributes(permission, permission.getTitle(), permission.getDescription(),
+				permission.getParameter());
 	}
 
 	@Override
-	public void changePermissionParameter(Permission permission) {
-		permissionDao.saveOrUpdate(permission);		
-	}	
-	
-	@Override
-	public void changePermissionAttributes(Permission perm, String newTitle, String newDescription, String newParamter) {
+	public void changePermissionAttributes(Permission perm, String newTitle, String newDescription,
+			String newParamter) {
+		if (!securityManager.isPermitted("cmrAdministrationPermission")) {
+			return;
+		}
+		
 		perm.setTitle(newTitle);
 		perm.setDescription(newDescription);
 		perm.setParameter(newParamter);
-		
+
 		permissionDao.saveOrUpdate(perm);
 	}
-	
+
 	@Override
 	public List<Permission> getAllPermissions() {
+		if (!securityManager.isPermitted("cmrAdministrationPermission")) {
+			return new ArrayList<Permission>();
+		}
+		
 		return permissionDao.loadAll();
 	}
 
 	// | ROLE | --------------
 	@Override
 	public Role getRoleByID(long id) throws DataRetrievalFailureException, DataIntegrityViolationException {
-		Role roles = roleDao.findByID(id);
+		if (!securityManager.isPermitted("cmrAdministrationPermission")) {
+			return null;
+		}
 		
+		Role roles = roleDao.findByID(id);
+
 		if (roles == null) {
-			throw new DataRetrievalFailureException("No roles in the database matching the given id!");			
+			throw new DataRetrievalFailureException("No roles in the database matching the given id!");
 		} else {
 			return roles;
 		}
@@ -295,22 +388,35 @@ public class SecurityService implements ISecurityService {
 
 	@Override
 	public Role getRoleOfUser(String email) throws AuthenticationException, DataIntegrityViolationException {
-		User foundUser = userDao.findByEmail(email);
+		if (!securityManager.isPermitted("cmrAdministrationPermission")) {
+			return null;
+		}
 		
+		User foundUser = userDao.findByEmail(email);
+
 		if (foundUser == null) {
-			throw new DataRetrievalFailureException("No user in the database matching the given email!");			
+			throw new DataRetrievalFailureException("No user in the database matching the given email!");
 		} else {
 			return getRoleByID(foundUser.getRoleId());
-		}		
+		}
 	}
 
 	@Override
 	public List<Role> getAllRoles() {
+		if (!securityManager.isPermitted("cmrAdministrationPermission")) {
+			return new ArrayList<Role>();
+		}
+		
 		return roleDao.loadAll();
 	}
 
 	@Override
-	public void addRole(String name, List<String> rolePermissions, String description) throws DataIntegrityViolationException {
+	public void addRole(String name, List<String> rolePermissions, String description)
+			throws DataIntegrityViolationException {
+		if (!securityManager.isPermitted("cmrAdministrationPermission")) {
+			return;
+		}
+		
 		List<Permission> allPermissions = getAllPermissions();
 		List<Permission> grantedPermissions = new ArrayList<Permission>();
 		for (int i = 0; i < rolePermissions.size(); i++) {
@@ -321,34 +427,120 @@ public class SecurityService implements ISecurityService {
 				}
 			}
 		}
-		Role role = new Role(name, grantedPermissions, description);		
-			
+		Role role = new Role(name, grantedPermissions, description);
+
 		if (!checkDataIntegrity(role)) {
 			throw new DataIntegrityViolationException("Data integrity test failed!");
 		}
 		List<Role> allRole = roleDao.loadAll();
 		if (allRole.contains(role)) {
 			throw new DataIntegrityViolationException("Role already exist!");
-		} else {			
+		} else {
 			roleDao.saveOrUpdate(role);
 		}
 	}
-	
+
 	@Override
-	public void changeRoleDescription(Role role, String newDescription)	{
+	public void changeRoleDescription(Role role, String newDescription) {
+		if (!securityManager.isPermitted("cmrAdministrationPermission")) {
+			return;
+		}
+		
 		changeRoleAttribute(role, role.getTitle(), newDescription, role.getPermissions());
 	}
-	
+
 	@Override
-	public void changeRoleAttribute(Role role, String newTitle, String newDescription, List<Permission> newPermissions) {
+	public void changeRoleAttribute(Role role, String newTitle, String newDescription,
+			List<Permission> newPermissions) {
+		if (!securityManager.isPermitted("cmrAdministrationPermission")) {
+			return;
+		}
+		
 		role.setTitle(newTitle);
 		role.setDescription(newDescription);
 		role.setPermissions(newPermissions);
 		roleDao.saveOrUpdate(role);
 	}
-	
+
 	@Override
 	public void deleteRole(Role role) {
+		if (!securityManager.isPermitted("cmrAdministrationPermission")) {
+			return;
+		}
+		
 		roleDao.delete(role);
+	}
+
+	@Override
+	public void resetDB() {
+		if (!securityManager.isPermitted("cmrAdministrationPermission")) {
+			return;
+		}
+
+		/**
+		 * All users logout.
+		 */
+		DefaultSessionManager sm = (DefaultSessionManager) securityManager.getSessionManager();
+
+		for (Session session : sm.getSessionDAO().getActiveSessions()) {
+			new Subject.Builder().session(session).buildSubject().logout();
+		}
+
+		userDao.deleteAll(userDao.loadAll());
+		roleDao.deleteAll(roleDao.loadAll());
+		permissionDao.deleteAll(permissionDao.loadAll());
+
+		/**
+		 * copied from info.novatec.inspectit.cmr.security.SecurityInitialization start()
+		 */
+		if (permissionDao.loadAll().isEmpty()) {
+
+			Permission cmrRecordingPermission = new Permission("cmrRecordingPermission",
+					"Permission to start recording from Agent");
+			Permission cmrShutdownAndRestartPermission = new Permission("cmrShutdownAndRestartPermission",
+					"Permission for shutting down and restarting the CMR");
+			Permission cmrDeleteAgentPermission = new Permission("cmrDeleteAgentPermission",
+					"Permission for deleting an Agent");
+			Permission cmrStoragePermission = new Permission("cmrStoragePermission",
+					"Permission for accessing basic storage options");
+			Permission cmrAdministrationPermission = new Permission("cmrAdministrationPermission",
+					"Permission for accessing the CMR Administration");
+			Permission cmrLookAtAgentsPermission = new Permission("cmrLookAtAgentsPermission",
+					"General permission to look at agents.");
+
+			// Transfers permissions to database.
+			permissionDao.saveOrUpdate(cmrRecordingPermission);
+			permissionDao.saveOrUpdate(cmrShutdownAndRestartPermission);
+			permissionDao.saveOrUpdate(cmrDeleteAgentPermission);
+			permissionDao.saveOrUpdate(cmrStoragePermission);
+			permissionDao.saveOrUpdate(cmrAdministrationPermission);
+			permissionDao.saveOrUpdate(cmrLookAtAgentsPermission);
+
+			// Predefined roles
+			Role guestRole = new Role("guestRole", new ArrayList<Permission>(), "The role of a guest-user.");
+			Role restrictedRole = new Role("restrictedRole",
+					Arrays.asList(cmrRecordingPermission, cmrStoragePermission, cmrLookAtAgentsPermission),
+					"The role of a restricted-user.");
+			Role adminRole = new Role("adminRole",
+					Arrays.asList(cmrRecordingPermission, cmrStoragePermission, cmrDeleteAgentPermission,
+							cmrShutdownAndRestartPermission, cmrAdministrationPermission, cmrLookAtAgentsPermission),
+					"The role of an admin-user.");
+
+			// Transfers roles to database.
+			roleDao.saveOrUpdate(guestRole);
+			roleDao.saveOrUpdate(restrictedRole);
+			roleDao.saveOrUpdate(adminRole);
+
+			// Standarduser - has to be changed on first login
+			User admin = new User(Permutation.hashString("admin"), "admin", adminRole.getId(), false);
+
+			// Guestuser - can be edited to give a user without an account
+			// rights
+			User guest = new User(Permutation.hashString("guest"), "guest", guestRole.getId(), false);
+
+			// Transfers users to databse.
+			userDao.saveOrUpdate(guest);
+			userDao.saveOrUpdate(admin);
+		}
 	}
 }
